fix: properly set available outputs when not waiting for run end
Motivation:

All output are currently set only if 'waitForRunEnd' is true; but some of them can be set as soon as the run has started.

Modifications:

Set the outputs 'run_id', 'reports_url', and 'runs_url', even if 'waitForRunEnd' is false.

Result:

The outputs 'run_id', 'reports_url', and 'runs_url' are always available if the run was started successfully.