Handle interruptions in the docker runner
If the container is stopped with 'docker stop', or in case of error in this runner, while a run is ongoing and we are waiting for the run end, try to cancel the run.

Note: this does not seem to work when interrupting a GitLab job. We could add a command to be called in 'after_script', but this is currently isn't executed when the job is cancelled... Might be worth revisiting when this is implemented: https://gitlab.com/gitlab-org/gitlab/-/issues/387230